pyproject.toml
[build-system]
requires = [
	"setuptools >= 61",
	"setuptools_scm",
	"wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "lyscripts"
description = "Package to interact with lymphatic progression data and models."
authors = [
	{name = "Roman Ludwig", email = "roman.ludwig@usz.ch"}
]
readme = "README.md"
requires-python = ">=3.10"
keywords = ["scripts", "lymph", "inference"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy",
    "pandas",
    "emcee",
    "h5py",
    "tables",
    "matplotlib",
    "corner",
    "scipy",
    "xlrd",
    "rich",
    "rich-argparse",
    "pyyaml",
    "lymph-model >= 1.2.3",
    "deprecated",
    "joblib",
    "pydantic",
    "pydantic-settings >= 2.7.0",
    "numpydantic",
    "lydata >= 0.2.0",
    "loguru",
]
dynamic = ["version"]

[project.urls]
source = "https://github.com/rmnldwg/lyscripts"
documentation = "https://lyscripts.readthedocs.io"

[project.optional-dependencies]
docs = [
    "sphinx",
    "sphinx-book-theme",
    "sphinx-autodoc-typehints",
    "sphinxcontrib-programoutput",
    "myst_parser",
    "autodoc_pydantic",
]
test = [
    "pytest",
    "pytest-mpl",
]

[project.scripts]
lyscripts = "lyscripts:main"

[tool.setuptools.packages.find]
include = ["lyscripts"]

[tool.setuptools_scm]
write_to = "lyscripts/_version.py"
local_scheme = "no-local-version"

[tool.setuptools.dynamic]
version = {attr = "lyscripts._version.version"}

[tool.pytest.ini_options]
testpaths = "."

[tool.ruff.lint]
select = ["E", "F", "W", "B", "C", "R", "U", "D", "I", "S", "T", "A", "N"]
ignore = ["D409"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs}/*" = [
    "D103",
    "E402",
    "S101",
    "S603",
    "S607",
]


# git-cliff ~ default configuration file
# https://git-cliff.org/docs/configuration
#
# Lines starting with "#" are comments.
# Configuration options are organized into tables and keys.
# See documentation for more information on available options.

[tool.git-cliff.changelog]
# changelog header
header = """
# Changelog\n
All notable changes to this project will be documented in this file.\n
"""
# template for the changelog body
# https://tera.netlify.app/docs
body = """
{% if version %}\
    ## [{{ version | trim_start_matches(pat="v") }}] - {{ timestamp | date(format="%Y-%m-%d") }}
{% else %}\
    ## [unreleased]
{% endif %}\
{% for group, commits in commits | group_by(attribute="group") %}
    ### {{ group | upper_first }}
    {% for c in commits %}
        - {% if c.breaking %}⚠ **BREAKING** {% endif -%}
        {% if c.scope %}(**{{ c.scope }}**) {% endif -%}
        {{ c.message | upper_first }}.
        {%- if c.footers %}{% for f in c.footers %}{% if not f.breaking %} {{ f.token }} [{{ f.value }}].{% endif %}{% endfor %}{% endif %}
        {%- if c.body %}\\
        {{ c.body | indent(prefix="  ", first=True) }}
        {% endif -%}
    {% endfor %}
{% endfor %}\n
"""
# remove the leading and trailing whitespace from the template
trim = true
# changelog footer
footer = """
<!-- generated by git-cliff -->
"""

[tool.git-cliff.git]
# parse the commits based on https://www.conventionalcommits.org
conventional_commits = true
# filter out the commits that are not conventional
filter_unconventional = true
# process each line of a commit as an individual commit
split_commits = false
# regex for preprocessing the commit messages
commit_preprocessors = [
  # { pattern = '\((\w+\s)?#([0-9]+)\)', replace = "([#${2}](https://github.com/orhun/git-cliff/issues/${2}))"}, # replace issue numbers
]
# regex for parsing and grouping commits
commit_parsers = [
  { message = "^feat", group = "Features" },
  { message = "^fix", group = "Bug Fixes" },
  { message = "^docs", group = "Documentation" },
  { message = "^perf", group = "Performance" },
  { message = "^refactor", group = "Refactor" },
  { message = "^style", group = "Styling" },
  { message = "^test", group = "Testing" },
  { message = "^chore\\(release\\): prepare for", skip = true },
  { message = "^chore", group = "Miscellaneous Tasks" },
  { body = ".*security", group = "Security" },
]
# protect breaking changes from being skipped due to matching a skipping commit_parser
protect_breaking_commits = false
# filter out the commits that are not matched by commit parsers
filter_commits = false
# glob pattern for matching git tags
tag_pattern = "[0-9]*"
# regex for skipping tags
skip_tags = "v0.1.0-beta.1"
# regex for ignoring tags
ignore_tags = ""
# sort the tags topologically
topo_order = false
# sort the commits inside sections by oldest/newest order
sort_commits = "oldest"
# limit the number of commits included in the changelog.
# limit_commits = 42
